refactor: securityConfig
