+++ src/main/java/com/bovo/Bovo/config/securityConfig.java
@@ -41,17 +41,18 @@ public class securityConfig {
                 .csrf(AbstractHttpConfigurer::disable) // CSRF 보호 기능 비활성화
                 .formLogin(AbstractHttpConfigurer::disable) // 기본 로그인 페이지 비활성화
                 .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS)) // 세션 비활성화
-                .authorizeHttpRequests(auth -> { // 요청(URL)에 대한 접근 제어 설정
-                    auth.requestMatchers("/**", "/error", "/static/**").permitAll(); // 개발 중 임시 인증 없이 허용
-//                    auth.requestMatchers(HttpMethod.OPTIONS, "/**").permitAll(); // Preflight 요청 허용
-//                    auth.requestMatchers(HttpMethod.POST,"/","/login", "/my-page/logout", "/register", "/register/nickname", "/register/email", "/refresh").permitAll(); // 로그인과 회원가입 인증 없이 허용
-//                    auth.requestMatchers(HttpMethod.GET, "/**").permitAll();
-//                    auth.requestMatchers(HttpMethod.DELETE, "/**").authenticated();
-//                    auth.requestMatchers(HttpMethod.PUT, "/**").authenticated();
-//                    auth.requestMatchers(HttpMethod.POST, "/**").authenticated(); // 위 요청 외에는 인증 필요
-                })
-//                .addFilterBefore(jwtFilter, UsernamePasswordAuthenticationFilter.class) // Jwt 인증 필터 추가
                 .cors(cors -> cors.configurationSource(corsConfig()))
+                .authorizeHttpRequests(auth -> { // 요청(URL)에 대한 접근 제어 설정
+//                    auth.requestMatchers("/**", "/error", "/static/**").permitAll(); // 개발 중 임시 인증 없이 허용
+                    auth.requestMatchers(HttpMethod.OPTIONS, "/**").permitAll(); // Preflight 요청 허용
+                    auth.requestMatchers(HttpMethod.POST,"/","/login", "/my-page/logout", "/register", "/register/nickname", "/register/email", "/refresh").permitAll(); // 로그인과 회원가입 인증 없이 허용
+                    auth.requestMatchers(HttpMethod.GET, "/my-page/**").authenticated();
+                    auth.requestMatchers(HttpMethod.GET, "/**").permitAll();
+                    auth.requestMatchers(HttpMethod.DELETE, "/**").authenticated();
+                    auth.requestMatchers(HttpMethod.PUT, "/**").authenticated();
+                    auth.requestMatchers(HttpMethod.POST, "/**").authenticated(); // 위 요청 외에는 인증 필요
+                })
+                .addFilterBefore(jwtFilter, UsernamePasswordAuthenticationFilter.class) // Jwt 인증 필터 추가
                 .build();
 
     }
